refactor(cli): consolidate config resolution
