+++ packages/cli/src/index.ts
@@ -2,10 +2,10 @@ export { cdn } from 'vite-plugin-monkey';
 export type {
 	AppConfig,
 	CliConfig,
+	MakooConfig as MakooOptions,
 	MonkeyBuildConfig,
 	MonkeyConfig,
 	MonkeyServerConfig
 } from './config/types';
 export { makoo } from './vite/makoo';
 export { makooDev } from './vite/makooDev';
-export type { MakooOptions } from './vite/types';
