+++ packages/cli/src/vite/makoo.ts
@@ -1,10 +1,9 @@
 import type { Plugin } from 'vite';
 import monkey from 'vite-plugin-monkey';
-import { resolveConfig } from '../config/resolve';
-import { resolveMonkeyPluginOptions } from './toMonkeyOptions';
-import type { MakooOptions } from './types';
+import { resolveConfig, resolveMonkeyPluginOptions } from '../config/resolve';
+import type { MakooConfig } from '../config/types';
 
-export function makoo(options: MakooOptions): Plugin[] {
+export function makoo(options: MakooConfig): Plugin[] {
 	const { root, ...config } = options;
 	const resolvedConfig = resolveConfig(config, root);
 	const monkeyOptions = resolveMonkeyPluginOptions(resolvedConfig);
