--- packages/cli/src/vite/types.ts
@@ -1,5 +1,0 @@
-import type { CliConfig } from '../config/types';
-
-export type MakooOptions = CliConfig & {
-	root?: string;
-};
