+++ packages/cli/test/config/resolve.test.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { describe, expect, it } from 'vitest';
-import { resolveConfig } from '../../src/config/resolve';
+import { resolveConfig, resolveMonkeyPluginOptions } from '../../src/config/resolve';
 
 const root = path.resolve('/project');
 
@@ -31,5 +31,33 @@ describe('resolveConfig', () => {
 		expect(config.monkey.server.mountGmApi).toBe(false);
 		expect(config.monkey.build.fileName).toBe('demo-script.user.js');
 		expect(config.monkey.build.metaFileName).toBe('demo-script.meta.js');
+	});
+
+	it('resolves options for vite-plugin-monkey', () => {
+		const config = resolveConfig(
+			{
+				entry: './src/main.ts',
+				app: { name: 'demo-script', version: '1.2.3' },
+				monkey: {
+					userscript: {
+						namespace: 'https://makoo.test',
+						match: ['https://example.com/*']
+					},
+					server: { open: false }
+				}
+			},
+			root
+		);
+
+		const options = resolveMonkeyPluginOptions(config);
+
+		expect(options.entry).toBe(path.join(root, 'src/main.ts'));
+		expect(options.userscript).toMatchObject({
+			name: { '': 'demo-script' },
+			version: '1.2.3',
+			namespace: 'https://makoo.test'
+		});
+		expect(options).not.toHaveProperty('clientAlias');
+		expect(options.server).toMatchObject({ open: false, mountGmApi: false });
 	});
 });
